+++ test/context.go
@@ -30,10 +30,8 @@ import (
 
 // WithLogger allows using a Logger created with a MockWriter quickly.
 func WithLogger(t *testing.T, f func(logger *xylog.Logger, w *MockWriter)) {
-	xylog.SetBufferSize(1)
-	defer xylog.SetBufferSize(4096)
 	var writer = &MockWriter{}
-	var emitter = xylog.NewStreamEmitter(writer)
+	var emitter = xylog.NewStreamEmitter(writer, 1)
 	var handler = xylog.GetHandler("")
 	handler.AddEmitter(emitter)
 
@@ -48,10 +46,8 @@ func WithLogger(t *testing.T, f func(logger *xylog.Logger, w *MockWriter)) {
 
 // WithHandler allows using a Handler with MockWriter.
 func WithHandler(t *testing.T, f func(h *xylog.Handler, w *MockWriter)) {
-	xylog.SetBufferSize(1)
-	defer xylog.SetBufferSize(4096)
 	var writer = &MockWriter{}
-	var emitter = xylog.NewStreamEmitter(writer)
+	var emitter = xylog.NewStreamEmitter(writer, 1)
 	var handler = xylog.GetHandler(t.Name())
 	handler.AddEmitter(emitter)
 
@@ -60,19 +56,15 @@ func WithHandler(t *testing.T, f func(h *xylog.Handler, w *MockWriter)) {
 
 // WithStreamEmitter allows using a StreamEmitter created with a MockWriter
 // quickly.
-func WithStreamEmitter(
-	t *testing.T, f func(e *xylog.StreamEmitter, w *MockWriter),
-) {
-	xylog.SetBufferSize(1)
-	defer xylog.SetBufferSize(4096)
+func WithStreamEmitter(t *testing.T, f func(e *xylog.StreamEmitter, w *MockWriter)) {
 	var writer = &MockWriter{}
-	var emitter = xylog.NewStreamEmitter(writer)
+	var emitter = xylog.NewStreamEmitter(writer, 1)
 	f(emitter, writer)
 }
 
 // WithBenchLogger allows using a Logger whose output is io.Discard.
 func WithBenchLogger(b *testing.B, f func(logger *xylog.Logger)) {
-	var emitter = xylog.NewStreamEmitter(io.Discard)
+	var emitter = xylog.NewStreamEmitter(io.Discard, 4096)
 	var handler = xylog.GetHandler("")
 	handler.AddEmitter(emitter)
 	handler.SetEncoding(encoding.NewJSONEncoding())